// src/data/property-types.ts
export type Property = {
  slug: string;
  title: string;
  location: string;
  locationTag: string;
  price: number;
  priceFormatted: string;
  description: string;
  bedrooms: number;
  bathrooms: number;
  area: string;
  type: string;
  image: string;
  gallery: string[];
  features: string[];
};